// src/services/igdb/generated/types.gen.ts
export type Checksum = string;
// ...
export type Cover = {
    id?: Id;
    alpha_channel?: boolean;
    animated?: boolean;
    game?: Id;
    /**
     * Reference ID for Game Localization this cover might be associated with
     */
    game_localization?: string;
    /**
     * The height of the image in pixels
     */
    height?: number;
    /**
     * The ID of the image used to construct an IGDB image link
     */
    image_id?: string;
    /**
     * The website address (URL) of the item
     */
    url?: string;
    /**
     * The width of the image in pixels
     */
    width?: number;
    checksum?: Checksum;
};

/**
 * Date this was initially added to the IGDB database
 */
export type CreatedAt = number;
// ...
export type ExternalGame = {
    id?: Id;
    category?: ExternalGameCategoryEnums;
    external_game_source?: Id;
    /**
     * The ISO country code of the external game product.
     */
    countries?: Array<number>;
    created_at?: CreatedAt;
    game?: Id;
    media?: ExternalGameMediaEnums;
    game_release_format?: Id;
    /**
     * The name of the game according to the other service
     */
    name?: string;
    platform?: Id;
    /**
     * The other services ID for this game
     */
    uid?: string;
    updated_at?: UpdatedAt;
    /**
     * The website address (URL) of the item
     */
    url?: string;
    /**
     * The year in full (2018)
     */
    year?: number;
    checksum?: Checksum;
};
// ...
export type ExternalGameCategoryEnums = 1 | 5 | 10 | 11 | 13 | 14 | 15 | 20 | 22 | 23 | 26 | 28 | 29 | 30 | 31 | 32 | 36 | 37 | 54 | 55;

/**
 * The type of media
 * 1: digital
 * 2: physical
 *
 *
 * @deprecated
 */
export type ExternalGameMediaEnums = 1 | 2;
// ...
export type GameTimeToBeat = {
    id?: Id;
    game_id?: Id;
    /**
     * Average time (in seconds) to finish the game to its credits without spending notable time on extras such as side quests.
     */
    hastily?: number;
    /**
     * Average time (in seconds) to finish the game while mixing in some extras such as side quests without being overly thorough.
     */
    normally?: number;
    /**
     * Average time (in seconds) to finish the game to 100% completion.
     */
    completely?: number;
    /**
     * Total number of time to beat submissions for this game
     */
    count?: number;
    created_at?: CreatedAt;
    updated_at?: UpdatedAt;
    checksum?: Checksum;
};
// ...
export type Genre = {
    id?: Id;
    created_at?: CreatedAt;
    /**
     * The name of the genre
     */
    name?: string;
    /**
     * A url-safe, unique, lower-case version of the name
     */
    slug?: string;
    updated_at?: UpdatedAt;
    /**
     * The website address (URL) of the item
     */
    url?: string;
    checksum?: Checksum;
};
// ...
export type Id = number;
// ...
export type Screenshot = {
    id?: Id;
    alpha_channel?: boolean;
    animated?: boolean;
    game?: Id;
    /**
     * The height of the image in pixels
     */
    height?: number;
    /**
     * The ID of the image used to construct an IGDB image link
     */
    image_id?: string;
    /**
     * The website address (URL) of the item
     */
    url?: string;
    /**
     * The width of the image in pixels
     */
    width?: number;
    checksum?: Checksum;
};
// ...
export type UpdatedAt = number;